refactor: move login up to not logout on message load
