+++ src/pages/HomePage/HomePage.tsx
@@ -12,16 +12,8 @@ import { SubmenuSubItem } from '../../components/Submenu/SubmenuSubItem/SubmenuS
 import { HeaderUserIcons } from '../../components/HeaderUserIcons/HeaderUserIcons';
 import { Submenu } from '../../components/Submenu/Submenu';
 import { MainSection } from '../../components/MainSection/MainSection';
-import {
-  GoogleLogin,
-  GoogleLoginResponse,
-  GoogleLoginResponseOffline,
-  GoogleLogout,
-} from 'react-google-login';
 import { useUserContext } from '../../hooks/useUserContext';
-import { useUserData } from '../../hooks/useUserData';
 import {
-  clientId,
   emptyProfilePictureSrc,
   SubItemCategories,
   categoryContainer,
@@ -29,24 +21,6 @@ import {
 } from '../../components/config';
 export const UserContext = useUserContext;
 export const HomePage: React.FC = () => {
-  const { state, dispatch } = useUserData();
-
-  const responseGoogle = (
-    response: GoogleLoginResponse | GoogleLoginResponseOffline,
-  ) => {
-    if ('profileObj' in response) {
-      dispatch({ type: 'logIn', user: response });
-    }
-  };
-  const responseFailure = (resp: { error: string }) => {
-    console.log(resp);
-    // TODO: add error handler
-    throw new Error(resp.error);
-  };
-  const logout = () => {
-    dispatch({ type: 'logOut' });
-  };
-
   const subItems = SubItemCategories.map((subItemProps) => {
     return (
       <SubmenuSubItem
@@ -54,7 +28,6 @@ export const HomePage: React.FC = () => {
         href={subItemProps.href}
         name={subItemProps.name}
         category={subItemProps.category}
-        // isActive={false}
       >
         {subItemProps.name}
       </SubmenuSubItem>
@@ -75,44 +48,27 @@ export const HomePage: React.FC = () => {
 
   return (
     <section>
-      <UserContext.Provider value={state.userData}>
-        <GoogleLogin
-          clientId={clientId}
-          buttonText="Login"
-          onSuccess={responseGoogle}
-          onFailure={responseFailure}
-          disabled={state.isLogged}
-          disabledStyle={{ display: 'none' }}
-        />
-        <GoogleLogout
-          clientId={clientId}
-          buttonText="Logout"
-          onLogoutSuccess={logout}
-          disabled={!state.isLogged}
-          disabledStyle={{ display: 'none' }}
-        />
-        <Head>
-          <HeaderLogo
-            mobileImgSrc={testProps_logo.mobileImgSrc}
-            mobileImgAlc={testProps_logo.mobileImgAlc}
-            desktopImgSrc={testProps_logo.desktopImgSrc}
-            desktopImgAlt={testProps_logo.desktopImgAlt}
-          >
-            Page Title
-          </HeaderLogo>
-          <HeaderUser src={emptyProfilePictureSrc}>
-            <HeaderUserIcons classname={'icon-bell'} />
-            <HeaderUserIcons classname={'icon-mail'} />
-          </HeaderUser>
-        </Head>
-        <Navbar>{navBarItems}</Navbar>
-        <MainContainer>
-          <Submenu>
-            <SubmenuItem name={categoryContainer.name}>{subItems}</SubmenuItem>
-          </Submenu>
-          <MainSection />
-        </MainContainer>
-      </UserContext.Provider>
+      <Head>
+        <HeaderLogo
+          mobileImgSrc={testProps_logo.mobileImgSrc}
+          mobileImgAlc={testProps_logo.mobileImgAlc}
+          desktopImgSrc={testProps_logo.desktopImgSrc}
+          desktopImgAlt={testProps_logo.desktopImgAlt}
+        >
+          Page Title
+        </HeaderLogo>
+        <HeaderUser src={emptyProfilePictureSrc}>
+          <HeaderUserIcons classname={'icon-bell'} />
+          <HeaderUserIcons classname={'icon-mail'} />
+        </HeaderUser>
+      </Head>
+      <Navbar>{navBarItems}</Navbar>
+      <MainContainer>
+        <Submenu>
+          <SubmenuItem name={categoryContainer.name}>{subItems}</SubmenuItem>
+        </Submenu>
+        <MainSection />
+      </MainContainer>
     </section>
   );
 };
